+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changes since v2.1
 - Update some examples in `examples/marine`.
 - Fix an bug reported by Ken Mankoff: scripts in `examples/antarctica/` required PISM
   built with PROJ.
+- Implementation of a Climate Index approach for glacial/interglacial simulations
 
 Changes since v2.0
 ==================